% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{analyze}
\alias{analyze}
\title{Categorize Function}
\usage{
analyze(path)
}
\arguments{
\item{path}{The path relative to working directory that holds data in
designated "categorized" directories.}
}
\value{
A data frame with two columns composed of (image_id, category)
tuples.
}
\description{
\code{analyze} loads counted categories and determines the best category for
each image.
}
\examples{

}
