% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_htg_index.R
\name{calculate_htg_index}
\alias{calculate_htg_index}
\title{Calculate heterogeneity index for category responses}
\usage{
calculate_htg_index(row)
}
\arguments{
\item{row}{A list of counts by category for a given image.}
}
\value{
A measure of variance relative to the heterogeneity of repsonses.
}
\description{
\code{calculate_htg_index} calculates a measure of heterogeneity
for a given image's category response counts in a single row.
}
