% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorize.R
\name{categorize}
\alias{categorize}
\title{Categorize Function}
\usage{
categorize(path, filename = "categorized.tsv")
}
\arguments{
\item{path}{The path relative to working directory that holds data in
designated "valid" and "keys" directories.}

\item{filename}{The name of the output file that will be written after
categorization is finished.}
}
\description{
\code{categorize} loads, pivots, and combines valid blocks
of image categorization task responses, then associates
participant response counts to categories for each
image.
}
\examples{

}
