% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{clean}
\alias{clean}
\title{Clean Function}
\usage{
clean(path)
}
\arguments{
\item{path}{The path relative to working directory that holds data in
designated "raw" directory.}
}
\description{
\code{clean} scrubs a sequenctial list of Qualtrics data export files
detailing participant category responses to images within a directory of
image blocks.
}
\examples{
clean("../mounts/imgct/data/5_category")

}
