% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_qualtrics_export.R,
%   R/parse_qualtrics_export.R
\name{clean_qualtrics_export}
\alias{clean_qualtrics_export}
\title{Parse Qualtrics export survey response data}
\usage{
clean_qualtrics_export(
  path,
  export_name = "qualtrics_export",
  qualtrics_tag = "_Q10"
)

clean_qualtrics_export(
  path,
  export_name = "qualtrics_export",
  qualtrics_tag = "_Q10"
)
}
\arguments{
\item{path}{A path to data directories for project.}

\item{export_name}{The name of the exported file from Qualtrics. Note: the
file must be a TSV, and the export options must prepare values as numeric
that were specially coded in the study.}

\item{qualtrics_tag}{A string literal matching the qualtrics tag that was assigned to naming by Qualtrics that
should be removed for analysis. Defaults to "_Q10" for now.}
}
\value{
Returns parsed data frame established prior to looping and
writing individualized response files.

Returns message of success.
}
\description{
\code{clean_qualtrics_export} removes artifacts that are embedded in the raw data exports from Qualtrics.

\code{clean_qualtrics_export} removes artifacts that are embedded in the raw data exports from Qualtrics.
}
\examples{
clean_qualtrics_export(path, export_name)

clean_qualtrics_export(path, export_name)

}
