% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorize.R
\name{.count_categories}
\alias{.count_categories}
\title{\code{.count_categories} uses a combination of long and wide pivots to
provide counted category responses for a given image block.}
\usage{
.count_categories(block_df, cat_id_list)
}
\arguments{
\item{block_df}{A dataframe from loading a clean, validated set of responses
to images.}

\item{cat_id_list}{A list of the identiers used for categorization in the
given task.}
}
\description{
\code{.count_categories} uses a combination of long and wide pivots to
provide counted category responses for a given image block.
}
