% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorize.R
\name{.write}
\alias{.write}
\title{\code{.write} writes a dataframe to the absolute path composed
of path and filename.}
\usage{
.write(df, path, filename = "categorized.tsv")
}
\arguments{
\item{df}{The dataframe to write to file.}

\item{path}{The path to the directory in which the file will be written.}

\item{filename}{The filename of the file that will be written. Defaults to
"categorized.tsv".}
}
\description{
\code{.write} writes a dataframe to the absolute path composed
of path and filename.
}
