% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_all_clean_blocks.R
\name{load_all_clean_blocks}
\alias{load_all_clean_blocks}
\title{Load All Clean Image Block Responses}
\usage{
load_all_clean_blocks(path)
}
\arguments{
\item{path}{The path string to the directory in which clean data lives.}
}
\value{
all_clean_blocks A list of dataframes of the clean blocks. 
The dataframes will have NA values replaced by 0.
}
\description{
\code{load_all_clean_block} loads all clean blocks at a specified full path to the
directory in which they all exist with extension TSV.
}
\examples{
load_all_clean_block("path/to/clean")

}
