% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_all_participant_validations.R
\name{load_all_participant_validations}
\alias{load_all_participant_validations}
\title{Load All Participant Validations}
\usage{
load_all_participant_validations(path)
}
\arguments{
\item{path}{The path string to the parent directory in which the validation
file exists. This should be the same parent directory for "clean" block data.}
}
\value{
all_participant_validations A dataframe consisting of all
participants and all validation responses with a total number of correct
validations.
}
\description{
\code{load_all_participant_validations} loads the TSV at specified full path to the
the "results" subdirectory of the desired data path.
}
\examples{
load_all_participant_validations(path)

}
