% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_block.R
\name{load_block}
\alias{load_block}
\title{Load block from file}
\usage{
load_block(path, block_template = c("RAW", "CLEAN", "VALID"))
}
\arguments{
\item{path}{String for relative path.}

\item{block_template}{String literal for block template.}
}
\value{

}
\description{
\code{load_block} reads a block into a dataframe.
}
\examples{
load_block("./data/raw/raw_01.tsv", block_template = "RAW")
load_block("./data/clean/clean_07.tsv", block_template = "CLEAN")

}
