% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_clean_block.R
\name{load_clean_block}
\alias{load_clean_block}
\title{Load Clean Image Block Responses}
\usage{
load_clean_block(path)
}
\arguments{
\item{path}{The path string to the clean block to load.}
}
\value{
clean_block A dataframe of the clean block.
}
\description{
\code{load_clean_block}
}
\examples{
load_clean_block(path)

}
