% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_imported_xlsx.R
\name{load_imported_xlsx}
\alias{load_imported_xlsx}
\title{Load Imported XLSX}
\usage{
load_imported_xlsx(path, import_filename, columns = NULL)
}
\arguments{
\item{path}{The path string to the parent directory in which the all data
files exists.}

\item{import_filename}{The filename of the result desired. Defaults to
XLSX, and no XLSX extension is required, but is accepted. Other file
extensions are not accepted.}

\item{columns}{Types of columns desired when loading imported dataset.
Defaults to guesses by readxl interpreter. Override default guesses with 
a vector of column types, such as c("guess", "text", "guess", "skip").}
}
\value{
result A dataframe consisting of the imported database file's contents.
}
\description{
\code{load_imported_xlsx} loads the .xlsx file at specified full path to the
the "imports" subdirectory of the desired data path.
}
\examples{
load_imported_xlsx(path, "iaps_emotion_ratings")

}
