% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_key.R
\name{load_key}
\alias{load_key}
\title{Load Key from TXT file}
\usage{
load_key(path, filename, col_names = c("key", "value"))
}
\arguments{
\item{path}{String path to the project root directory.}

\item{filename}{Name of the file with extension .txt.}

\item{col_names}{Vector of the key and value headers. Defaults to "key"
and "value" via c("key", "value").}
}
\value{
A transposed key.
}
\description{
\code{load_key} reads a list of key=value pairs in a txt file where "=" is
the required delimeter.
}
\examples{
load_key("/path/to/keys", filename = "validation_key.txt")
load_key("/path/to/keys", filename = "categorization_key.txt")

}
