% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_qualtrics_export.R
\name{load_qualtrics_export}
\alias{load_qualtrics_export}
\title{Load Qualtrics Export (TSV)}
\usage{
load_qualtrics_export(path, filename = "qualtrics_export.tsv")
}
\arguments{
\item{path}{The path string to the qualtrics export to load.}

\item{filename}{The string name of the TSV file, including the extension,
that the exported data was saved with at the  directory of the path
provided. Defaults to "qualtrics_export.tsv".}
}
\value{
qualtrics_export A dataframe of the raw exported Qualtrics data without
any processing.
}
\description{
\code{load_qualtrics_export}
}
\examples{
load_qualtrics_export("/path/to/qualtrics_export")

}
