% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_qualtrics_tsv.R
\name{load_qualtrics_tsv}
\alias{load_qualtrics_tsv}
\title{Load Qualtrics Data Export (TSV)}
\usage{
load_qualtrics_tsv(path, filename = "qualtrics.tsv")
}
\arguments{
\item{path}{The path string to the qualtrics export to load.}

\item{filename}{The string name of the TSV file, including the extension,
that the exported data was saved with at the  directory of the path
provided. Defaults to "qualtrics_export.tsv".}
}
\value{
qualtrics_df A dataframe of the raw exported Qualtrics data without
any processing.
}
\description{
\code{load_qualtrics_tsv} Load an exported dataset from Qualtrics that is in
TSV format.
}
\examples{
load_qualtrics_tsv("/path/to/qualtrics_export")

}
