% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_result.R
\name{load_result}
\alias{load_result}
\title{Load Result}
\usage{
load_result(
  path,
  result_filename,
  columns = readr::cols(.default = readr::col_character())
)
}
\arguments{
\item{path}{The path string to the parent directory in which the validation
file exists. This should be the same parent directory for "clean" block data.}

\item{result_filename}{The filename of the result desired. Defaults to
a TSV, and no TSV extension is required, but is accepted. Other file
extensions are not accepted.}

\item{columns}{Types of columns desired when loading a result dataset.
Defaults to all character types. Override default with a readr-style cols
definition.}
}
\value{
result A dataframe consisting of the result file's contents.
}
\description{
\code{load_result} loads the file at specified full path to the
the "results" subdirectory of the desired data path.
}
\examples{
load_result(path)

}
