% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_txt_key.R
\name{load_txt_key}
\alias{load_txt_key}
\title{Load a key from txt file}
\usage{
load_txt_key(
  path,
  key_template = c("CATEGORIZATION", "VALIDATION"),
  col_names = c("key", "value")
)
}
\arguments{
\item{path}{String path to file, which must be a relative path to the project root directory}

\item{key_template}{String literal that can equal "CATEGORIZATION" or "VALIDATION".}

\item{col_names}{Vector of column names for txt key. Defaults to c("key", "value").}
}
\value{
A transposed key.
}
\description{
\code{load_txt_key} reads a list of key=value pairs in a txt file.
}
\examples{
load_txt_key("./keys/validation_key.txt", key_template = "VALIDATION")

}
