% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_image_block.R
\name{make_image_blocks_from_dirs}
\alias{make_image_blocks_from_dirs}
\title{Make Image Blocks From Dirs Function}
\usage{
make_image_blocks_from_dirs(dbs_path, validation_path, output_path, block_size)
}
\arguments{
\item{dbs_path}{The path relative to working directory of database lists of
images.}

\item{validation_path}{The path relative to working directory of validation
image files.}

\item{output_path}{The path relative to working directory of output block
TXT files.}

\item{block_size}{The number of images per block.}
}
\value{
List of image file names with prefix database identifiers.
}
\description{
\code{make_image_blocks_from_dirs} Creates and writes a TXT blocks of image names.
   Notably, this function uses a list of lists of image files from
   databases to construct an equally but randomly allocated list of images
   from each in a series of equally sized blocks.
}
\examples{
make_image_list_from_dirs("imgct/images/databases",
 "imgct/images/validation",
 "imgct/images/blocks",
 300)

}
