% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_image_list.R
\name{make_image_list_from_dirs}
\alias{make_image_list_from_dirs}
\title{Make Image List Function}
\usage{
make_image_list_from_dirs(path, list_name = "img_list")
}
\arguments{
\item{path}{The path relative to working directory where database
images are stored in multiple directories, each named for the database.}

\item{list_name}{String name of list for output file. Defaults to
"image_list".}
}
\value{
List of image file names with prefix database identifiers.
}
\description{
\code{make_image_list} Creates and writes a TXT list of image names.
}
\examples{
make_image_list_from_dirs("imgct/images/databases")

}
