% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_qualtrics_export.R
\name{parse_qualtrics_export}
\alias{parse_qualtrics_export}
\title{Parse Qualtrics export survey response data}
\usage{
parse_qualtrics_export(
  path_to_export,
  path_to_blocks,
  path_to_raw,
  qualtrics_tag = "_Q10"
)
}
\arguments{
\item{path_to_export}{A path to the exported survey response data from qualtrics.}

\item{path_to_blocks}{A path to the directory containing ordered image block lists}

\item{path_to_raw}{A path to the directory containing individual raw
response blocks}

\item{qualtrics_tag}{A string literal matching the qualtrics tag that was assigned to naming by Qualtrics that
should be removed for analysis. Defaults to "_Q10" for now.}
}
\value{
Returns message of success; saves parsed and mutated raw data blocks
with correct image name headers.
}
\description{
\code{parse_qualtrics_export} removes artifacts that are embedded in the raw data exports from Qualtrics.
}
\examples{
parse_qualtrics_export(path_to_export, path_to_blocks, path_to_raw)

}
