% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_qualtrics_artifacts.R
\name{remove_qualtrics_artifacts}
\alias{remove_qualtrics_artifacts}
\title{Remove Qualtrics artifacts from a}
\usage{
remove_qualtrics_artifacts(block, block_id)
}
\description{
Remove Qualtrics artifacts from a
}
