% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_qualtrics_artifacts.R
\name{sub_qualtrics_image_string}
\alias{sub_qualtrics_image_string}
\title{Scrub qualtrics image string at end of name}
\usage{
sub_qualtrics_image_string(str)
}
\description{
Scrub qualtrics image string at end of name
}
