% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate}
\alias{validate}
\title{Validate Function}
\usage{
validate(path, threshold = 0, remove_below_threshold = FALSE)
}
\arguments{
\item{path}{The path relative to working directory that holds data in
designated "clean" and "keys" directories.}

\item{threshold}{The minimum level of validation accuracy for a given
participant required to be considered a "valid" point for further analysis. 
Defaults to 0, meaning no threshold.}

\item{remove_below_threshold}{Boolean determining whether participants
below validation threshold are removed from output. Defaults to false.}
}
\description{
\code{validate} validates a sequence of scrubbed (clean) blocks of data
against a validation key.
}
\examples{
validate("../mounts/imgct/data/5_category")

}
