% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate_all_participants}
\alias{validate_all_participants}
\title{Validate All Participant Responses In All Blocks}
\usage{
validate_all_participants(path, validation_key)
}
\arguments{
\item{path}{String path}

\item{validation_key}{Key dataframe}
}
\description{
\code{validate_all_participants}
}
