% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_block_participants.R
\name{validate_block_participants}
\alias{validate_block_participants}
\title{Validate All Participant Responses In One Block}
\usage{
validate_block_participants(clean_block, validation_key)
}
\arguments{
\item{clean_block}{A loaded dataframe of participant responses 
to a specified block of images. The dataframe must be 
scrubbed of Qualtrics or other artifacts prior to usage with this function.}

\item{validation_key}{A loaded validation key.}
}
\value{
participant_validations The block is returned as a validation block
consisting of only validation images and participant responses, plus a count
of valid responses per participant.
}
\description{
\code{validate_block_participants} assesses the validation image responses
by participants in a given clean image block. The validations are made
relative to a validation key, and the function returns a dataframe of
validation responses and total number of validations correctly made by
participant.
}
\examples{
validate_block_participants(clean_block_03, validation_key)

}
