% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_block.R
\name{load_block}
\alias{load_block}
\title{Load block from file}
\usage{
load_block(path, block_template = c("RAW", "CLEAN", "VALID"))
}
\arguments{
\item{block_template}{}
}
\value{

}
\description{
\code{load_block} reads a block into a dataframe.
}
\examples{

}
