% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_clean_block.R
\name{load_clean_block}
\alias{load_clean_block}
\title{Load Clean Image Block Responses}
\usage{
load_clean_block(path)
}
\arguments{
\item{path}{The path string to the clean block to load.}
}
\value{
clean_block A dataframe of the clean block. The dataframe will have
NA values replaced by 0.
}
\description{
\code{load_clean_block} loads a clean block at a specified full path to the
filename with extension, replaces NA values with 0, and returns the clean
block as a dataframe. Note that the data are loaded as character type across
all columns.
}
\examples{
load_clean_block(path)

}
