% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_qualtrics_export.R
\name{parse_qualtrics_export}
\alias{parse_qualtrics_export}
\title{Parse Qualtrics Export Dataframe}
\usage{
parse_qualtrics_export(
  qualtrics_export,
  qualtrics_tag = "_Q10",
  drop_rows = c(2),
  remove_participant_code = TRUE
)
}
\arguments{
\item{qualtrics_export}{The path string to the clean block to load.}

\item{qualtrics_tag}{String literal for Qualtrics naming convention needed
for parsing. Defaults to "_Q10".}

\item{drop_rows}{A vector of row numbers to drop. Defaults to `c(2)`.}

\item{remove_participant_code}{A boolean to remove the MTurk-created
participantCode. Useful if wanting to obfuscate participants. Defaults to
TRUE, but change to FALSE if needing to reference raw codes in analysis.}
}
\value{
parsed_qualtrics_export A dataframe of the parsed qualtrics export
data, meaning no artifacts are left behind.
}
\description{
\code{parse_qualtrics_export}
}
\examples{
parse_qualtrics_export(qualtrics_export, remove_participant_code = TRUE)

}
