% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_qualtrics_artifacts.R
\name{remove_qualtrics_artifacts}
\alias{remove_qualtrics_artifacts}
\title{Remove Qualtrics artifacts from a}
\usage{
remove_qualtrics_artifacts(block)
}
\arguments{
\item{block}{A dataframe of a block of loaded raw data from qualtrics.}
}
\value{
Return block as dataframe without qualtrics artifacts.
}
\description{
\code{remove_qualtrics_artifacts} removes artifacts that are embedded in the raw data exports from Qualtrics.
}
\examples{

}
