% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_all_participants.R
\name{validate_all_participants}
\alias{validate_all_participants}
\title{Validate All Participant Responses In All Blocks}
\usage{
validate_all_participants(path, key_filename = "validation_key.txt")
}
\arguments{
\item{path}{String path to the directory containing clean block responses,
keys and output for validated participant data.}

\item{key_filename}{Name of the validation key TXT file with extension.
Defaults to "validation_key.txt".}
}
\value{
A single master validation dataframe with participants and their
validation responses, as well as a count of valid responses.
}
\description{
\code{validate_all_participants} validates a sequence of scrubbed (clean) blocks of 
participant response data against a validation key. The
}
\examples{
validate_all_participants("/path/to/data", "validation_key.txt")

}
